Make sure correct type is used when calling `retroarch_get_flags()`

+++ tasks/task_content.c
@@ -1913,7 +1913,7 @@ bool task_push_start_dummy_core(content_ctx_info_t *content_info)
    const char *path_dir_system             = settings->paths.directory_system;
    bool check_firmware_before_loading      = settings->bools.check_firmware_before_loading;
 #ifdef HAVE_PATCH
-   uint16_t rarch_flags                    = retroarch_get_flags();
+   uint32_t rarch_flags                    = retroarch_get_flags();
 #endif
 
    if (!content_info)
@@ -2014,7 +2014,7 @@ bool task_push_load_content_from_playlist_from_menu(
 #endif
    bool check_firmware_before_loading         = settings->bools.check_firmware_before_loading;
 #ifdef HAVE_PATCH
-   uint16_t rarch_flags                       = retroarch_get_flags();
+   uint32_t rarch_flags                       = retroarch_get_flags();
 #endif
 
    content_ctx.flags     = 0;
@@ -2166,7 +2166,7 @@ bool task_push_start_current_core(content_ctx_info_t *content_info)
       content_ctx.flags |= CONTENT_INFO_FLAG_CHECK_FW_BEFORE_LOADING;
 #ifdef HAVE_PATCH
    {
-      uint16_t rarch_flags = retroarch_get_flags();
+      uint32_t rarch_flags = retroarch_get_flags();
       if (rarch_flags & RARCH_FLAGS_IPS_PREF)
          content_ctx.flags |= CONTENT_INFO_FLAG_IS_IPS_PREF;
       if (rarch_flags & RARCH_FLAGS_BPS_PREF)
@@ -2412,7 +2412,7 @@ bool task_push_load_content_with_new_core_from_menu(
       content_ctx.flags |= CONTENT_INFO_FLAG_CHECK_FW_BEFORE_LOADING;
 #ifdef HAVE_PATCH
    {
-      uint16_t rarch_flags  = retroarch_get_flags();
+      uint32_t rarch_flags  = retroarch_get_flags();
       if (rarch_flags & RARCH_FLAGS_IPS_PREF)
          content_ctx.flags |= CONTENT_INFO_FLAG_IS_IPS_PREF;
       if (rarch_flags & RARCH_FLAGS_BPS_PREF)
@@ -2522,7 +2522,7 @@ static bool task_load_content_internal(
    const char *path_dir_system             = settings->paths.directory_system;
    const char *path_dir_cache              = settings->paths.directory_cache;
 #ifdef HAVE_PATCH
-   uint16_t rarch_flags                    = retroarch_get_flags();
+   uint32_t rarch_flags                    = retroarch_get_flags();
 #endif
    content_ctx.flags                       = 0;
 
@@ -3025,7 +3025,7 @@ bool content_init(void)
    const char *path_dir_system        = settings->paths.directory_system;
    const char *path_dir_cache         = settings->paths.directory_cache;
 #ifdef HAVE_PATCH
-   uint16_t rarch_flags               = retroarch_get_flags();
+   uint32_t rarch_flags               = retroarch_get_flags();
 #endif
 
    content_file_list_free(p_content->content_list);
